[Fix][LinearAttention] Check every declared input, and count what it moves
The seven ops this branch promoted took their declared inputs on trust.
The saved forward artifacts — DeltaNet's S/Aw/Au/w/u, Gated DeltaNet's S,
GLA's h/dht — reached the kernel with no dtype, shape or device check, and
the ops compared dtypes inline instead of through the one gate the design
names. Each now calls the generated _validate_dtypes with its full input
list, and check_tensor_shape covers the shapes and the device.

GLAFwdOp returned (o, None) unless constructed with output_final_state,
which an entry with two declared outputs cannot do. The kernel computes the
state either way, so the switch is gone and the state always comes back.
The op also never handed initial_state to the kernel, and the entry typed
it same_as(q), which no caller could satisfy while feeding back an fp32
final_state; it is fp32 now, so one call's state seeds the next.

The rooflines counted the tensors the pre-promotion benchmarks counted,
which predate those outputs: DeltaNet forward omitted the five buffers it
returns, GLA forward omitted g and the final state, and the backwards
omitted the states they read. They now count what the ops move, fp32 state
at four bytes and the rest at the input dtype.

+++ workloads/linear_attention.py
@@ -299,9 +299,10 @@ class GLAChunkwiseWorkload(WorkloadBase):
         k = torch.randn(B, T, H, K, device="cuda", dtype=self.dtype) * 0.1
         v = torch.randn(B, T, H, V, device="cuda", dtype=self.dtype) * 0.1
         g = -torch.rand(B, T, H, K, device="cuda", dtype=self.dtype)
-        # Absent means None: the recurrence then starts from zeros.
+        # Absent means None: the recurrence then starts from zeros. Present, it is
+        # fp32, the dtype the recurrence carries the state in.
         initial_state = (
-            torch.randn(B, H, K, V, device="cuda", dtype=self.dtype) * 0.1
+            torch.randn(B, H, K, V, device="cuda", dtype=torch.float32) * 0.1
             if self.has_initial_state
             else None
         )
